created navigation namespace
